+++ docs/bindings_http_receiver_binary.js.html
@@ -8,7 +8,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
   <title>
-      formats/json/parser.js - Documentation
+      bindings/http/receiver_binary.js - Documentation
   </title>
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.7.0/highlight.min.js"></script>
@@ -81,7 +81,7 @@
   <div id="main">
     
       <h1 class="page-title">
-        formats/json/parser.js
+        bindings/http/receiver_binary.js
       </h1>
     
 
@@ -90,43 +90,60 @@
 
 <section>
   <article>
-    <pre class="prettyprint source linenums"><code>const {
-  isString,
-  isDefinedOrThrow,
-  isStringOrObjectOrThrow
-} = require("../../bindings/http/validation/fun.js");
-const ValidationError = require("../../bindings/http/validation/validation_error.js");
+    <pre class="prettyprint source linenums"><code>const ReceiverV1 = require("./v1/receiver_binary_1.js");
+const ReceiverV3 = require("./v03/receiver_binary_0_3.js");
 
-const invalidPayloadTypeError = new ValidationError("invalid payload type, allowed are: string or object");
-const nullOrUndefinedPayload = new ValidationError("null or undefined payload");
+const { SPEC_V03, SPEC_V1 } = require("./constants.js");
+const { check, parse } = require("./validation/binary.js");
 
-const asJSON = (v) => (isString(v) ? JSON.parse(v) : v);
-
-class JSONParser {
-  constructor(decorator) {
-    this.decorator = decorator;
+/**
+ * A class that receives binary CloudEvents over HTTP. This class can be used
+ * if you know that all incoming events will be using binary transport. If
+ * events can come as either binary or structured, use {HTTPReceiver}.
+ */
+class BinaryHTTPReceiver {
+  /**
+   * Creates a new BinaryHTTPReceiver to accept events over HTTP.
+   *
+   * @param {string} version the Cloud Event specification version to use. Default "1.0"
+   */
+  constructor(version = SPEC_V1) {
+    if (version === SPEC_V1) {
+      this.receiver = new ReceiverV1();
+    } else if (version === SPEC_V03) {
+      this.receiver = new ReceiverV3();
+    }
   }
 
   /**
-   * Parses the payload with an optional decorator
-   * @param {object|string} payload the JSON payload
-   * @return {object} the parsed JSON payload.
+   * Checks an incoming HTTP request to determine if it conforms to the
+   * Cloud Event specification for this receiver.
+   *
+   * @throws {ValidationError} if the event does not conform to the spec
+   * @param {Object} payload the HTTP request body
+   * @param {Object} headers  the HTTP request headers
+   *
+   * @returns {void}
    */
-  parse(payload) {
-    if (this.decorator) {
-      payload = this.decorator.parse(payload);
-    }
+  check(payload, headers) {
+    return check(payload, headers, this.receiver);
+  }
 
-    return Array.of(payload)
-
-      .filter((p) => isDefinedOrThrow(p, nullOrUndefinedPayload))
-      .filter((p) => isStringOrObjectOrThrow(p, invalidPayloadTypeError))
-      .map(asJSON)
-      .shift();
+  /**
+   * Parses an incoming HTTP request, converting it to a {CloudEvent}
+   * instance if it conforms to the Cloud Event specification for this receiver.
+   *
+   * @throws {ValidationError} of the event does not conform to the spec
+   * @param {Object} payload the HTTP request body
+   * @param {Object} headers the HTTP request headers
+   * @returns {CloudEvent} aninstance of CloudEvent representing the incoming request
+   */
+  parse(payload, headers) {
+    return parse(payload, headers, this.receiver);
   }
 }
 
-module.exports = JSONParser;
+module.exports = BinaryHTTPReceiver;
 </code></pre>
   </article>
 </section>
