+++ lib/bindings/http/emitter_binary.js
@@ -1,4 +1,6 @@
 const axios = require("axios");
+const EmitterV1 = require("./v1").BinaryEmitter;
+const EmitterV3 = require("./v03").BinaryEmitter;
 
 const {
   HEADERS,
@@ -32,10 +34,10 @@ class BinaryHTTPEmitter {
    */
   constructor(version) {
     if (version === SPEC_V1) {
-      this.headerByGetter = require("./emitter_binary_1.js");
+      this.headerByGetter = EmitterV1;
       this.extensionPrefix = BINARY_HEADERS_1.EXTENSIONS_PREFIX;
     } else if (version === SPEC_V03) {
-      this.headerByGetter = require("./emitter_binary_0_3.js");
+      this.headerByGetter = EmitterV3;
       this.extensionPrefix = BINARY_HEADERS_03.EXTENSIONS_PREFIX;
     }
   }
