+++ lib/bindings/http/http_receiver.js
@@ -1,8 +1,10 @@
-const V03Binary = require("./receiver_binary_0_3.js");
-const V03Structured = require("./receiver_structured_0_3.js");
-const V1Binary = require("./receiver_binary_1.js");
-const V1Structured = require("./receiver_structured_1.js");
-const ValidationError = require("../../validation_error.js");
+// const V03Binary = require("./receiver_binary_0_3.js");
+// const V03Structured = require("./v03/receiver_structured_0_3.js");
+// const V1Binary = require("./receiver_binary_1.js");
+// const V1Structured = require("./v1/receiver_structured_1.js");
+const BinaryReceiver = require("./receiver_binary.js");
+const StructuredReceiver = require("./receiver_structured.js");
+const ValidationError = require("./validation/validation_error.js");
 const {
   BINARY,
   STRUCTURED,
@@ -24,12 +26,12 @@ class HTTPReceiver {
   constructor() {
     this.receivers = {
       v1: {
-        structured: new V1Structured(),
-        binary: new V1Binary()
+        structured: new StructuredReceiver(SPEC_V1),
+        binary: new BinaryReceiver(SPEC_V1)
       },
       v03: {
-        structured: new V03Structured(),
-        binary: new V03Binary()
+        structured: new StructuredReceiver(SPEC_V03),
+        binary: new BinaryReceiver(SPEC_V03)
       }
     };
   }
