+++ lib/bindings/http/receiver_binary.js
@@ -1,134 +1,54 @@
-const { HEADER_CONTENT_TYPE, MIME_JSON, DEFAULT_SPEC_VERSION_HEADER } =
-  require("./constants.js");
-const Commons = require("./commons.js");
-const CloudEvent = require("../../cloudevent.js");
-const ValidationError = require("../../validation_error.js");
+const ReceiverV1 = require("./v1/receiver_binary_1.js");
+const ReceiverV3 = require("./v03/receiver_binary_0_3.js");
 
-const {
-  isDefinedOrThrow,
-  isStringOrObjectOrThrow
-} = require("../../utils/fun.js");
+const { SPEC_V03, SPEC_V1 } = require("./constants.js");
+const { check, parse } = require("./validation/binary.js");
 
-function validateArgs(payload, attributes) {
-  Array.of(payload)
-    .filter((p) => isDefinedOrThrow(p, new ValidationError("payload is null or undefined")))
-    .filter((p) => isStringOrObjectOrThrow(p, new ValidationError("payload must be an object or a string")))
-    .shift();
+/**
+ * A class that receives binary CloudEvents over HTTP. This class can be used
+ * if you know that all incoming events will be using binary transport. If
+ * events can come as either binary or structured, use {HTTPReceiver}.
+ */
+class BinaryHTTPReceiver {
+  /**
+   * Creates a new BinaryHTTPReceiver to accept events over HTTP.
+   *
+   * @param {string} version the Cloud Event specification version to use. Default "1.0"
+   */
+  constructor(version = SPEC_V1) {
+    if (version === SPEC_V1) {
+      this.receiver = new ReceiverV1();
+    } else if (version === SPEC_V03) {
+      this.receiver = new ReceiverV3();
+    }
+  }
 
-  Array.of(attributes)
-    .filter((a) => isDefinedOrThrow(a, new ValidationError("attributes is null or undefined")))
-    .shift();
+  /**
+   * Checks an incoming HTTP request to determine if it conforms to the
+   * Cloud Event specification for this receiver.
+   *
+   * @throws {ValidationError} if the event does not conform to the spec
+   * @param {Object} payload the HTTP request body
+   * @param {Object} headers  the HTTP request headers
+   *
+   * @returns {void}
+   */
+  check(payload, headers) {
+    return check(payload, headers, this.receiver);
+  }
+
+  /**
+   * Parses an incoming HTTP request, converting it to a {CloudEvent}
+   * instance if it conforms to the Cloud Event specification for this receiver.
+   *
+   * @throws {ValidationError} of the event does not conform to the spec
+   * @param {Object} payload the HTTP request body
+   * @param {Object} headers the HTTP request headers
+   * @returns {CloudEvent} aninstance of CloudEvent representing the incoming request
+   */
+  parse(payload, headers) {
+    return parse(payload, headers, this.receiver);
+  }
 }
-
-function BinaryHTTPReceiver(
-  parsersByEncoding,
-  setterByHeader,
-  allowedContentTypes,
-  requiredHeaders,
-  Spec,
-  specversion,
-  extensionsPrefix,
-  checkDecorator) {
-  this.parsersByEncoding = parsersByEncoding;
-  this.setterByHeader = setterByHeader;
-  this.allowedContentTypes = allowedContentTypes;
-  this.requiredHeaders = requiredHeaders;
-  this.Spec = Spec;
-  this.spec = new Spec();
-  this.specversion = specversion;
-  this.extensionsPrefix = extensionsPrefix;
-  this.checkDecorator = checkDecorator;
-}
-
-BinaryHTTPReceiver.prototype.check = function(payload, headers) {
-  // Validation Level 0
-  validateArgs(payload, headers);
-
-  if (this.checkDecorator) {
-    this.checkDecorator(payload, headers);
-  }
-
-  // Clone and low case all headers names
-  const sanityHeaders = Commons.sanityAndClone(headers);
-
-  // Validation Level 1 - if content-type exists, be sure it's
-  // an allowed type
-  const contentTypeHeader = sanityHeaders[HEADER_CONTENT_TYPE];
-  const noContentType = !this.allowedContentTypes.includes(contentTypeHeader);
-  if (contentTypeHeader && noContentType) {
-    throw new ValidationError("invalid content type", [sanityHeaders[HEADER_CONTENT_TYPE]]);
-  }
-
-  this.requiredHeaders
-    .filter((required) => !sanityHeaders[required])
-    .forEach((required) => {
-      throw new ValidationError(`header '${required}' not found`);
-    });
-
-  if (sanityHeaders[DEFAULT_SPEC_VERSION_HEADER] !==
-    this.specversion) {
-    throw new ValidationError("invalid spec version", [sanityHeaders[DEFAULT_SPEC_VERSION_HEADER]]);
-  }
-
-  // No erros! Its contains the minimum required attributes
-};
-
-function parserFor(parsersByEncoding, cloudevent, headers) {
-  const encoding = cloudevent.spec.payload.datacontentencoding;
-  return parsersByEncoding[encoding][headers[HEADER_CONTENT_TYPE]];
-}
-
-BinaryHTTPReceiver.prototype.parse = function(payload, headers) {
-  this.check(payload, headers);
-
-  // Clone and low case all headers names
-  const sanityHeaders = Commons.sanityAndClone(headers);
-  if (!sanityHeaders[HEADER_CONTENT_TYPE]) {
-    sanityHeaders[HEADER_CONTENT_TYPE] = MIME_JSON;
-  }
-
-  const processedHeaders = [];
-  const cloudevent = new CloudEvent(this.Spec);
-
-  // dont worry, check() have seen what was required or not
-  Array.from(Object.keys(this.setterByHeader))
-    .filter((header) => sanityHeaders[header])
-    .forEach((header) => {
-      const setterName = this.setterByHeader[header].name;
-      const parserFun = this.setterByHeader[header].parser;
-
-      // invoke the setter function
-      cloudevent[setterName](parserFun(sanityHeaders[header]));
-
-      // to use ahead, for extensions processing
-      processedHeaders.push(header);
-    });
-
-  // Parses the payload
-  const parsedPayload =
-      parserFor(this.parsersByEncoding, cloudevent, sanityHeaders)
-        .parse(payload);
-
-  // Every unprocessed header can be an extension
-  Array.from(Object.keys(sanityHeaders))
-    .filter((value) => !processedHeaders.includes(value))
-    .filter((value) =>
-      value.startsWith(this.extensionsPrefix))
-    .map((extension) =>
-      extension.substring(this.extensionsPrefix.length)
-    ).forEach((extension) =>
-      cloudevent.addExtension(extension,
-        sanityHeaders[this.extensionsPrefix + extension])
-    );
-
-  // Sets the data
-  cloudevent.data(parsedPayload);
-
-  // Checks the event spec
-  cloudevent.format();
-
-  // return the result
-  return cloudevent;
-};
 
 module.exports = BinaryHTTPReceiver;
