+++ lib/bindings/http/v03/emitter_binary_0_3.js
@@ -1,7 +1,7 @@
 const {
   HEADER_CONTENT_TYPE,
-  BINARY_HEADERS_1
-} = require("./constants.js");
+  BINARY_HEADERS_03
+} = require("../constants.js");
 
 const headerByGetter = {};
 
@@ -10,38 +10,43 @@ headerByGetter.getDataContentType = {
   parser: (v) => v
 };
 
+headerByGetter.getDataContentEncoding = {
+  name: BINARY_HEADERS_03.CONTENT_ENCODING,
+  parser: (v) => v
+};
+
 headerByGetter.getSubject = {
-  name: BINARY_HEADERS_1.SUBJECT,
+  name: BINARY_HEADERS_03.SUBJECT,
   parser: (v) => v
 };
 
 headerByGetter.getType = {
-  name: BINARY_HEADERS_1.TYPE,
+  name: BINARY_HEADERS_03.TYPE,
   parser: (v) => v
 };
 
 headerByGetter.getSpecversion = {
-  name: BINARY_HEADERS_1.SPEC_VERSION,
+  name: BINARY_HEADERS_03.SPEC_VERSION,
   parser: (v) => v
 };
 
 headerByGetter.getSource = {
-  name: BINARY_HEADERS_1.SOURCE,
+  name: BINARY_HEADERS_03.SOURCE,
   parser: (v) => v
 };
 
 headerByGetter.getId = {
-  name: BINARY_HEADERS_1.ID,
+  name: BINARY_HEADERS_03.ID,
   parser: (v) => v
 };
 
 headerByGetter.getTime = {
-  name: BINARY_HEADERS_1.TIME,
+  name: BINARY_HEADERS_03.TIME,
   parser: (v) => v
 };
 
-headerByGetter.getDataschema = {
-  name: BINARY_HEADERS_1.DATA_SCHEMA,
+headerByGetter.getSchemaurl = {
+  name: BINARY_HEADERS_03.SCHEMA_URL,
   parser: (v) => v
 };
