+++ lib/bindings/http/v03/index.js
@@ -1,0 +1,9 @@
+const Spec = require("./spec_0_3.js");
+const BinaryEmitter = require("./emitter_binary_0_3.js");
+const StructuredEmitter = require("./receiver_structured_0_3.js");
+
+module.exports = {
+  Spec,
+  BinaryEmitter,
+  StructuredEmitter
+};
