+++ lib/bindings/http/v03/receiver_structured_0_3.js
@@ -13,21 +13,19 @@ const {
     SUBJECT,
     DATA
   }
-} = require("./constants.js");
-const Constants = require("./constants.js");
-const Spec = require("../../specs/spec_0_3.js");
-const JSONParser = require("../../formats/json/parser.js");
-
-const StructuredHTTPReceiver = require("./receiver_structured.js");
+} = require("../constants.js");
+const Spec = require("./spec_0_3.js");
+const JSONParser = require("../../../formats/json/parser.js");
 
 const jsonParserSpec = new JSONParser();
+const parserByMime = {
+  [MIME_JSON]: jsonParserSpec,
+  [MIME_CE_JSON]: jsonParserSpec
+};
 
-const parserByMime = {};
-parserByMime[MIME_JSON] = jsonParserSpec;
-parserByMime[MIME_CE_JSON] = jsonParserSpec;
-
-const allowedContentTypes = [];
-allowedContentTypes.push(MIME_CE_JSON);
+const allowedContentTypes = [
+  MIME_CE_JSON
+];
 
 function parser(name, parser = (v) => v) {
   return { name: name, parser: parser};
@@ -46,23 +44,14 @@ parserMap.set(CONTENT_TYPE, passThroughParser("dataContentType"));
 parserMap.set(SUBJECT, passThroughParser("subject"));
 parserMap.set(DATA, passThroughParser("data"));
 
-// Leaving this in place for now. TODO: fixme
-// eslint-disable-next-line
-function Receiver(configuration) {
-  this.receiver = new StructuredHTTPReceiver(
-    parserByMime,
-    parserMap,
-    allowedContentTypes,
-    Spec
-  );
+class Receiver {
+  constructor() {
+    this.parserByMime = parserByMime;
+    this.parserMap = parserMap;
+    this.allowedContentTypes = allowedContentTypes;
+    this.Spec = Spec;
+    this.spec = new Spec();
+  }
 }
-
-Receiver.prototype.check = function(payload, headers) {
-  this.receiver.check(payload, headers);
-};
-
-Receiver.prototype.parse = function(payload, headers) {
-  return this.receiver.parse(payload, headers);
-};
 
 module.exports = Receiver;
