+++ lib/bindings/http/v03/spec_0_3.js
@@ -1,12 +1,12 @@
 const { v4: uuidv4 } = require("uuid");
 const Ajv = require("ajv");
-const ValidationError = require("../validation_error.js");
+const ValidationError = require("../validation/validation_error.js");
 
 const {
   isBase64,
   clone,
   asData
-} = require("../utils/fun.js");
+} = require("../validation/fun.js");
 
 const RESERVED_ATTRIBUTES = {
   type: "type",
@@ -126,7 +126,7 @@ function Spec03(_caller) {
   };
 
   if (!_caller) {
-    _caller = require("../cloudevent.js");
+    _caller = require("../../../cloudevent.js");
   }
 
   /*
