+++ lib/bindings/http/v1/emitter_binary_1.js
@@ -1,7 +1,7 @@
 const {
   HEADER_CONTENT_TYPE,
-  BINARY_HEADERS_03
-} = require("./constants.js");
+  BINARY_HEADERS_1
+} = require("../constants.js");
 
 const headerByGetter = {};
 
@@ -10,43 +10,38 @@ headerByGetter.getDataContentType = {
   parser: (v) => v
 };
 
-headerByGetter.getDataContentEncoding = {
-  name: BINARY_HEADERS_03.CONTENT_ENCODING,
-  parser: (v) => v
-};
-
 headerByGetter.getSubject = {
-  name: BINARY_HEADERS_03.SUBJECT,
+  name: BINARY_HEADERS_1.SUBJECT,
   parser: (v) => v
 };
 
 headerByGetter.getType = {
-  name: BINARY_HEADERS_03.TYPE,
+  name: BINARY_HEADERS_1.TYPE,
   parser: (v) => v
 };
 
 headerByGetter.getSpecversion = {
-  name: BINARY_HEADERS_03.SPEC_VERSION,
+  name: BINARY_HEADERS_1.SPEC_VERSION,
   parser: (v) => v
 };
 
 headerByGetter.getSource = {
-  name: BINARY_HEADERS_03.SOURCE,
+  name: BINARY_HEADERS_1.SOURCE,
   parser: (v) => v
 };
 
 headerByGetter.getId = {
-  name: BINARY_HEADERS_03.ID,
+  name: BINARY_HEADERS_1.ID,
   parser: (v) => v
 };
 
 headerByGetter.getTime = {
-  name: BINARY_HEADERS_03.TIME,
+  name: BINARY_HEADERS_1.TIME,
   parser: (v) => v
 };
 
-headerByGetter.getSchemaurl = {
-  name: BINARY_HEADERS_03.SCHEMA_URL,
+headerByGetter.getDataschema = {
+  name: BINARY_HEADERS_1.DATA_SCHEMA,
   parser: (v) => v
 };
