+++ lib/bindings/http/v1/index.js
@@ -1,0 +1,9 @@
+const Spec = require("./spec_1.js");
+const BinaryEmitter = require("./emitter_binary_1.js");
+const StructuredEmitter = require("./receiver_structured_1.js");
+
+module.exports = {
+  Spec,
+  BinaryEmitter,
+  StructuredEmitter
+};
