+++ lib/bindings/http/v1/receiver_structured_1.js
@@ -1,5 +1,4 @@
 const {
-  MIME_JSON,
   MIME_CE_JSON,
   STRUCTURED_ATTRS_1 : {
     TYPE,
@@ -11,23 +10,22 @@ const {
     CONTENT_TYPE,
     SUBJECT,
     DATA,
-    DATA_BASE64
+    DATA_BASE64,
+    MIME_JSON
   }
-} = require("./constants.js");
+} = require("../constants.js");
 
-const Spec = require("../../specs/spec_1.js");
-const JSONParser = require("../../formats/json/parser.js");
-
-const StructuredHTTPReceiver = require("./receiver_structured.js");
+const Spec = require("./spec_1.js");
+const JSONParser = require("../../../formats/json/parser.js");
 
 const jsonParserSpec = new JSONParser();
 
-const parserByMime = {};
-parserByMime[MIME_JSON] = jsonParserSpec;
-parserByMime[MIME_CE_JSON] = jsonParserSpec;
+const parserByMime = {
+  [MIME_JSON]: jsonParserSpec,
+  [MIME_CE_JSON]: jsonParserSpec
+};
 
-const allowedContentTypes = [];
-allowedContentTypes.push(MIME_CE_JSON);
+const allowedContentTypes = [ MIME_CE_JSON ];
 
 function parser(name, parser = (v) => v) {
   return { name: name, parser: parser};
@@ -46,23 +44,14 @@ parserMap.set(SUBJECT, passThroughParser("subject"));
 parserMap.set(DATA, passThroughParser("data"));
 parserMap.set(DATA_BASE64, passThroughParser("data"));
 
-// Leaving this in place for now. TODO: fixme
-// eslint-disable-next-line
-function Receiver(configuration) {
-  this.receiver = new StructuredHTTPReceiver(
-    parserByMime,
-    parserMap,
-    allowedContentTypes,
-    Spec
-  );
+class Receiver {
+  constructor() {
+    this.parserByMime = parserByMime;
+    this.parserMap = parserMap;
+    this.allowedContentTypes = allowedContentTypes;
+    this.Spec = Spec;
+    this.spec = new Spec();
+  }
 }
-
-Receiver.prototype.check = function(payload, headers) {
-  this.receiver.check(payload, headers);
-};
-
-Receiver.prototype.parse = function(payload, headers) {
-  return this.receiver.parse(payload, headers);
-};
 
 module.exports = Receiver;
