+++ lib/bindings/http/v1/spec_1.js
@@ -1,6 +1,6 @@
 const { v4: uuidv4 } = require("uuid");
 const Ajv = require("ajv");
-const ValidationError = require("../validation_error.js");
+const ValidationError = require("../validation/validation_error.js");
 
 const {
   asData,
@@ -10,7 +10,7 @@ const {
   isDate,
   isBinary,
   clone
-} = require("../utils/fun.js");
+} = require("../validation/fun.js");
 
 const isValidType = (v) =>
   (isBoolean(v) || isInteger(v) || isString(v) || isDate(v) || isBinary(v));
@@ -122,7 +122,7 @@ function Spec1(_caller) {
   };
 
   if (!_caller) {
-    _caller = require("../cloudevent.js");
+    _caller = require("../../../cloudevent.js");
   }
 
   /*
