+++ lib/bindings/http/validation/commons.js
@@ -1,4 +1,10 @@
-const Constants = require("./constants.js");
+const ValidationError = require("./validation_error.js");
+const Constants = require("../constants.js");
+const {
+  isDefinedOrThrow,
+  isStringOrObjectOrThrow
+} = require("./fun.js");
+
 
 // Specific sanity for content-type header
 function sanityContentType(contentType) {
@@ -27,7 +33,19 @@ function sanityAndClone(headers) {
   return sanityHeaders;
 }
 
+function validateArgs(payload, attributes) {
+  Array.of(payload)
+    .filter((p) => isDefinedOrThrow(p, new ValidationError("payload is null or undefined")))
+    .filter((p) => isStringOrObjectOrThrow(p, new ValidationError("payload must be an object or a string")))
+    .shift();
+
+  Array.of(attributes)
+    .filter((a) => isDefinedOrThrow(a, new ValidationError("attributes is null or undefined")))
+    .shift();
+}
+
 module.exports = {
   sanityAndClone,
-  sanityContentType
+  sanityContentType,
+  validateArgs
 };
