+++ lib/bindings/http/validation/validation_error.js
@@ -7,7 +7,7 @@ class ValidationError extends TypeError {
    * Constructs a new {ValidationError} with the message
    * and array of additional errors.
    * @param {string} message the error message
-   * @param {[string]|[ErrorObject]} [errors] any additional errors related to validation
+   * @param {string[]|ErrorObject[]} [errors] any additional errors related to validation
    */
   constructor(message, errors) {
     super(message);
