+++ lib/cloudevent.js
@@ -1,4 +1,4 @@
-const Spec = require("./specs/spec_1.js");
+const Spec = require("./bindings/http/v1/spec_1.js");
 const Formatter = require("./formats/json/formatter.js");
 
 /**
