+++ lib/formats/json/parser.js
@@ -2,8 +2,8 @@ const {
   isString,
   isDefinedOrThrow,
   isStringOrObjectOrThrow
-} = require("../../utils/fun.js");
-const ValidationError = require("../../validation_error.js");
+} = require("../../bindings/http/validation/fun.js");
+const ValidationError = require("../../bindings/http/validation/validation_error.js");
 
 const invalidPayloadTypeError = new ValidationError("invalid payload type, allowed are: string or object");
 const nullOrUndefinedPayload = new ValidationError("null or undefined payload");
