--- v03/index.js
@@ -1,5 +1,0 @@
-const Spec = require("../lib/specs/spec_0_3.js");
-
-module.exports = {
-  Spec
-};
