--- v1/index.js
@@ -1,5 +1,0 @@
-const Spec = require("../lib/specs/spec_1.js");
-
-module.exports = {
-  Spec
-};
